move argparse details to new file
